requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord-argparse==1.0.1
discord.py==1.5.1
docopt==0.6.2
idna==2.10
multidict==4.7.6
peewee==3.14.0
pipreqs==0.4.10
py-crypt==1.1.1
pycryptodome==3.9.9
pyDes==2.0.1
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.2
yarg==0.1.9
yarl==1.5.1
